requirements.txt
accelerate==0.31.0
asgiref==3.8.1
bitsandbytes==0.43.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
deskew==1.5.1
einops==0.8.0
filelock==3.15.4
flash-attn==2.5.9.post1
Flask==3.0.3
fsspec==2024.6.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
itsdangerous==2.2.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pillow==10.3.0
psutil==6.0.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pytesseract==0.3.10
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.14.0
sentencepiece==0.2.0
setuptools==70.1.1
sympy==1.12.1
tifffile==2024.6.18
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.2
typing_extensions==4.12.2
urllib3==2.2.2
waitress==3.0.0
Werkzeug==3.0.3
wheel==0.43.0
